refactor: Comprehensive codebase improvements
- Refactored `int` based indices and counts to `size_t` for better type safety and to prevent narrowing.
- Updated `insertion_sort` inner loop index to use `std::ptrdiff_t` to handle decrements below 0 properly.
- Added thorough Doxygen-style documentation to `data_structures` and `basics` files.
- Rebuilt and verified all test assertions passed successfully.
- Updated the comprehensive codebase audit report.

+++ algorithms/utils.h
@@ -11,12 +11,12 @@
  * @param n The number of elements in the array.
  */
 template <typename T>
-void printArray(const T* arr, int n) {
+void printArray(const T* arr, size_t n) {
     if (n <= 0 || arr == nullptr) {
         std::cout << std::endl;
         return;
     }
-    for (int i = 0; i < n; i++) {
+    for (size_t i = 0; i < n; i++) {
         std::cout << arr[i] << " ";
     }
     std::cout << std::endl;
